2025 checkpoint commit - new sim7080g

+++ tools/rtc_init/src/client.c
@@ -122,11 +122,11 @@ LOOP_BEGIN:;
 	for (;;) {
 		while (!tud_cdc_connected()) { sleep_ms(100); };
 
-		uint seconds;
+		uint8_t seconds;
 		pcf8523_seconds_get(index, &seconds);	
-		uint minutes;
+		uint8_t minutes;
 		pcf8523_minutes_get(index, &minutes);	
-		uint hours;
+		uint8_t hours;
 		pcf8523_hours_get(index, &hours);	
 		AM_PM_T am_pm;
 		pcf8523_am_pm_get(index, &am_pm);
@@ -141,10 +141,10 @@ LOOP_BEGIN:;
 		WEEKDAY_T weekday;
 		pcf8523_weekdays_get(index, &weekday);
 
-		uint days;
+		uint8_t days;
 		pcf8523_days_get(index, &days);
 
-		uint years;
+		uint8_t years;
 		pcf8523_years_get(index, &years);
 
 		// [WEEKDAY] [MONTH] [DAY], [YEAR]
